Initialize ocrWorker in UploadView to save startup-time in AttributeContainer
